Add "authorization" concept to the website
