+++ app/views/articles/edit.html.erb
@@ -1,4 +1,8 @@
 <h1><%= t("article.modify") %></h1>
+<small id="delete-link">
+  <span class="iconify" data-icon="ri:delete-bin-fill" data-inline="false"></span>
+  <%= link_to t("article.delete"), article_path(@article), method: :delete, data: { confirm: t("sure") } %>
+</small>
 <%= form_with model: @article, url: article_path(@article), method: :patch do |form| %>
   <%= render "form", model: @article, form: form, button: "article.modify" %>
 <% end %>
