+++ app/views/categories/show.html.erb
@@ -3,6 +3,6 @@
 <% else %>
   <p>
     <span class="iconify" data-icon="ri:error-warning-fill" data-inline="false"></span>
-    <%= t("user.no_articles") %>
+    <%= t("category.no_articles") %>
   </p>
 <% end %>
