+++ app/views/users/edit.html.erb
@@ -1,4 +1,8 @@
 <h1><%= t("user.edit_profile") %></h1>
+<small id="delete-link">
+  <span class="iconify" data-icon="ri:delete-bin-fill" data-inline="false"></span>
+  <%= link_to t("user.delete_profile"), user_path(@user), method: :delete, data: { confirm: t("sure") } %>
+</small>
 <%= form_with model: @user, url: user_path(@user), method: :patch do |form| %>
-  <%= render "form", model: @user, form: form, button: "user.edit" %>
+  <%= render "form", model: @user, form: form, button: "user.edit", link_path: nil %>
 <% end %>
