+++ app/views/users/sign_in.html.erb
@@ -1,4 +1,4 @@
 <h1><%= t("user.sign_in") %></h1>
 <%= form_with model: @user, url: sign_in_path do |form| %>
-  <%= render "form", model: @user, form: form, button: "user.sign_in" %>
+  <%= render "form", model: @user, form: form, button: "user.sign_in", link_text: "user.not_yet", link_path: sign_up_path %>
 <% end %>
