+++ app/views/users/sign_up.html.erb
@@ -1,4 +1,4 @@
 <h1><%= t("user.sign_up") %></h1>
 <%= form_with model: @user, url: sign_up_path do |form| %>
-  <%= render "form", model: @user, form: form, button: "user.sign_up" %>
+  <%= render "form", model: @user, form: form, button: "user.sign_up", link_text: "user.already", link_path: sign_in_path %>
 <% end %>
