+++ spec/models/user_spec.rb
@@ -1,4 +1,5 @@
 require "rails_helper"
+require "shared_rutines"
 
 RSpec.describe User, type: :model do
   context "validations" do
